requirements.txt
requests
sshtunnel
pika
toml
psutil
pexpect
pycryptodome
pyjwt
urllib3
PyQt5
tabulate
